feat:implemented pagination & sorting i Productmodule

+++ src/main/java/com/ecommerce/project/service/CategoryServiceImpl.java
@@ -52,7 +52,7 @@ public class CategoryServiceImpl implements CategoryService {
         categoryResponse.setContent(categoryDTOS);
         categoryResponse.setPageNumber(categorypage.getNumber());
         categoryResponse.setPageSize(categorypage.getSize());
-        categoryResponse.setTotalElements((int) categorypage.getTotalElements());
+        categoryResponse.setTotalElements(categorypage.getTotalElements());
         categoryResponse.setTotalPages(categorypage.getTotalPages());
         categoryResponse.setLastPage(categorypage.isLast());
